refactor: 보드와 관련된 상수의 위치 이동
