--- domain/src/main/kotlin/domain/Constants.kt
@@ -1,7 +1,0 @@
-package domain
-
-const val MIN_VIEW_X = 'A'
-const val MAX_VIEW_X = 'O'
-const val MIN_VIEW_Y = 1
-const val MAX_VIEW_Y = 15
-const val BOARD_SIZE = 15
