+++ domain/src/main/kotlin/view/BoardView.kt
@@ -1,6 +1,10 @@
 package view
 
 import domain.*
+import domain.Board.Companion.MAX_VIEW_X
+import domain.Board.Companion.MAX_VIEW_Y
+import domain.Board.Companion.MIN_VIEW_X
+import domain.Board.Companion.MIN_VIEW_Y
 import domain.stone.BlackStone
 import domain.stone.WhiteStone
