+++ domain/src/main/kotlin/view/OutputView.kt
@@ -1,8 +1,8 @@
 package view
 
 import domain.Board
-import domain.MIN_VIEW_X
-import domain.MIN_VIEW_Y
+import domain.Board.Companion.MIN_VIEW_X
+import domain.Board.Companion.MIN_VIEW_Y
 
 object OutputView {
